+++ CHANGELOG.md
@@ -8,8 +8,9 @@ Note that since we don't clearly distinguish between a public and private interf
 
 - [Breaking] Rename the ``model-index`` color theme to ``trajectory-index``
 - Add a new ``model-index`` color theme that uniquely colors each loaded model
-- Add the new ``model-index`` color theme as an option for the carbon color in the ``element-symbol`` and ``ilustrative`` color themes
-- Add nearest method to lookup3d
+- Add the new ``model-index`` and ``structure-index`` color themes as an option for the carbon color in the ``element-symbol`` and ``ilustrative`` color themes
+- Add ``structure-index`` color theme that uniquely colors each root structure
+- Add ``nearest`` method to ``Lookup3D``
 - Add mipmap-based blur for skybox backgrounds
 
 ## [v3.19.0] - 2022-10-01